Liferay upgrade 6.2 -> 7.0.
