+++ bnd.bnd
@@ -1,0 +1,9 @@
+Bundle-Name: advised-assets-portlet
+Bundle-SymbolicName: com.tuneit.advisedassets.portlet
+Bundle-Version: 2.0.0
+Export-Package: com.tuneit.advisedassets.constants
+-jsp: *.jsp,*.jspf
+-plugin.jsp: com.liferay.ant.bnd.jsp.JspAnalyzerPlugin
+-plugin.resourcebundle: com.liferay.ant.bnd.resource.bundle.ResourceBundleLoaderAnalyzerPlugin
+-plugin.sass: com.liferay.ant.bnd.sass.SassAnalyzerPlugin
+-sass: *
