+++ src/main/java/com/tuneit/advisedassets/constants/AdvisedAssetsPortletKeys.java
@@ -1,0 +1,10 @@
+package com.tuneit.advisedassets.constants;
+
+/**
+ * @author sky
+ */
+public class AdvisedAssetsPortletKeys {
+
+    public static final String AdvisedAssets = "advisedassets";
+
+}
