+++ src/main/java/com/tuneit/advisedassets/portlet/AdvisedAssetsPortlet.java
@@ -1,0 +1,28 @@
+package com.tuneit.advisedassets.portlet;
+
+import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
+import com.tuneit.advisedassets.constants.AdvisedAssetsPortletKeys;
+import org.osgi.service.component.annotations.Component;
+
+import javax.portlet.Portlet;
+
+/**
+ * @author sky
+ */
+@Component(
+    immediate = true,
+    property = {
+        "com.liferay.portlet.display-category=category.sample",
+        "com.liferay.portlet.instanceable=true",
+        "javax.portlet.display-name=AdvisedAssets",
+        "javax.portlet.name=" + AdvisedAssetsPortletKeys.AdvisedAssets,
+        "javax.portlet.security-role-ref=guest,power-user,user",
+        "javax.portlet.init-param.template-path=/",
+        "javax.portlet.init-param.view-template=/view.jsp",
+        "javax.portlet.resource-bundle=content.Language",
+        "com.liferay.portlet.css-class-wrapper=AdvisedAssets-portlet",
+        "javax.portlet.supports.mime-type=text/html"
+    },
+    service = Portlet.class
+)
+public class AdvisedAssetsPortlet extends MVCPortlet { }
