+++ src/main/resources/content/Language_en_US.properties
@@ -6,7 +6,7 @@ edit.maxassetslistsize=Max size of adviced assets list:
 edit.maxbannersnumber=Max number of banners:
 edit.minfavor=Min matching tags number:
 edit.bannerstagregexp=A regular expression that specifies the banners tags:
-edit.repeatbanner=Repeat resources that have a banner in the adviced assets list 
+edit.repeatbanner=Repeat resources that have a banner in the adviced assets list
 edit.save=Save
 
 javax.portlet.title=Adviced Content
