--- src/main/webapp/WEB-INF/liferay-hook.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE hook PUBLIC "-//Liferay//DTD Hook 6.2.0//EN" "http://www.liferay.com/dtd/liferay-hook_6_2_0.dtd">
-
-<hook>
-    <language-properties>
-        content/Language_en_US.properties
-    </language-properties>
-    <language-properties>
-        content/Language_ru.properties
-    </language-properties>
-    <custom-jsp-dir>/custom_jsps</custom-jsp-dir>
-</hook>
