+++ src/main/webapp/WEB-INF/liferay-plugin-package.properties
@@ -1,4 +1,4 @@
-name=AdvicedAssetsPortlet
+name=AdvisedAssetsPortlet
 module-group-id=liferay
 module-incremental-version=1
 tags=
@@ -7,5 +7,6 @@ change-log=
 page-url=http://www.tune-it.ru
 author=Tune IT, Ltd.
 licenses=LGPL
-liferay-versions=6.2.0+
+liferay-versions=7.0.0+
+portal-dependency-tlds=liferay-aui.tld,liferay-portlet.tld
 security-manager-enabled=false
