--- src/main/webapp/WEB-INF/liferay-portlet.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE liferay-portlet-app PUBLIC "-//Liferay//DTD Portlet Application 6.2.0//EN" "http://www.liferay.com/dtd/liferay-portlet-app_6_2_0.dtd">
-
-<liferay-portlet-app>
-    <portlet>
-        <portlet-name>AdvicedAssets</portlet-name>
-        <icon>/icon.png</icon>
-        <configuration-action-class>com.tuneit.advicedassets.AdvicedAssetsConfig</configuration-action-class>
-        <instanceable>false</instanceable>
-        <header-portlet-css>/css/main.css</header-portlet-css>
-        <footer-portlet-javascript>/js/main.js</footer-portlet-javascript>
-        <css-class-wrapper>AdvicedAssets-portlet</css-class-wrapper>
-    </portlet>
-    <role-mapper>
-        <role-name>administrator</role-name>
-        <role-link>Administrator</role-link>
-    </role-mapper>
-    <role-mapper>
-        <role-name>guest</role-name>
-        <role-link>Guest</role-link>
-    </role-mapper>
-    <role-mapper>
-        <role-name>power-user</role-name>
-        <role-link>Power User</role-link>
-    </role-mapper>
-    <role-mapper>
-        <role-name>user</role-name>
-        <role-link>User</role-link>
-    </role-mapper>
-</liferay-portlet-app>
