--- src/main/webapp/WEB-INF/portlet.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0"?>
-
-<portlet-app xmlns="http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd" version="2.0">
-    <portlet>
-        <portlet-name>AdvicedAssets</portlet-name>
-        <display-name>AdvicedAssets</display-name>
-        <portlet-class>com.liferay.util.bridges.mvc.MVCPortlet</portlet-class>
-        <init-param>
-            <name>view-template</name>
-            <value>/html/view.jsp</value>
-        </init-param>
-        <expiration-cache>0</expiration-cache>
-        <supports>
-            <mime-type>text/html</mime-type>
-        </supports>
-        <portlet-info>
-            <title>Adviced Content</title>
-            <short-title>Adviced Content</short-title>
-            <keywords>Adviced, Content</keywords>
-        </portlet-info>
-        <security-role-ref>
-            <role-name>administrator</role-name>
-        </security-role-ref>
-        <security-role-ref>
-            <role-name>guest</role-name>
-        </security-role-ref>
-        <security-role-ref>
-            <role-name>power-user</role-name>
-        </security-role-ref>
-        <security-role-ref>
-            <role-name>user</role-name>
-        </security-role-ref>
-    </portlet>
-</portlet-app>
