+++ src/main/webapp/WEB-INF/web.xml
@@ -1,18 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <web-app id="WebApp_ID" version="2.4" xmlns="http://java.sun.com/xml/ns/j2ee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee http://java.sun.com/xml/ns/j2ee/web-app_2_4.xsd">
-    <display-name>AdvicedAssets-portlet</display-name>
+    <display-name>AdvisedAssets-portlet</display-name>
 	
     <jsp-config>
         <taglib>
             <taglib-uri>http://java.sun.com/portlet_2_0</taglib-uri>
-            <taglib-location>
-                /WEB-INF/tld/liferay-portlet.tld
-            </taglib-location>
+            <taglib-location>/WEB-INF/tld/liferay-portlet.tld</taglib-location>
         </taglib>
         <taglib>
             <taglib-uri>http://liferay.com/tld/aui</taglib-uri>
-            <taglib-location>/WEB-INF/tld/aui.tld</taglib-location>
+            <taglib-location>/WEB-INF/tld/liferay-aui.tld</taglib-location>
         </taglib>
+
     </jsp-config>
 	
 </web-app>
